requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
APScheduler==3.10.4
attrs==24.2.0
discord.py==2.4.0
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
multidict==6.1.0
polib==1.2.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
wheel==0.44.0
yarl==1.13.1